feat: add schema migration script and initial schema
A common rough spot when multiple people are developing against the DB is for someone to change the schema, have it work locally, but not work on other developer boxes or in prod. An easy way to share these transitions from one schema version to the next is with a migrator. A lot of ORMs have them baked in. Since we aren't using an ORM, this utility provides just that.