feat: add heroku.properties
We have our JDBC variables in environment variables in Heroku. The heroku.properties will reference these ENV var names. Added logic to treat the values in the properties as ENV var names if the property `ricebay.env` is true. Finally, updated the deployment to set the system property `ricebay.environ` to "heroku" when deploying to Heroku.