Sorry for the massive amount of changes. I've accomplished most views but there are some things still needed to be done including:
CreateAuction: needs to connect with back end to create the auction. Needs to support image uploading. Validation is incredibly basic and needs more work.

EditAuction: needs to be created/auto-populated. The form follows CreateAuction but needs to populate the fields with existing auction details.

AuctionAction: I've created this as a way to implement the PlaceBid/RateSeller/ChangeRating button based on conditionals. I don't believe it can currently be implemented based on the existing auction/bid endpoints as not enough data is obtainable.

In a follow-up commit I will implement tests for purchase history and my auctions.